+++ database/factories/OrganisationFactory.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Database\Factories;
+
+use App\Models\Organisation;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+/**
+ * @extends Factory<Organisation>
+ */
+class OrganisationFactory extends Factory
+{
+    protected $model = Organisation::class;
+
+    public function definition(): array
+    {
+        return [
+            'name' => fake()->unique()->company(),
+            'type' => fake()->randomElement(array_keys(Organisation::TYPES)),
+            'parent_id' => null,
+            'notes' => null,
+        ];
+    }
+}