+++ resources/views/admin/location/edit.blade.php
@@ -19,6 +19,7 @@
 
             <div class="flex justify-end mt-8 gap-4">
                 @isset($location)
+                    <x-ui.button type="button" color="outline" href="{{ route('admin.workshop.history', ['location_id' => $location->id]) }}">Workshop history</x-ui.button>
                     <x-ui.button type="button" color="danger" x-data x-on:click.prevent="SM.confirmDelete('{{ csrf_token() }}', 'Delete Location?', 'Are you sure you want to delete this location? This action cannot be undone', '{{ route('admin.location.destroy', $location) }}')">Delete</x-ui.button>
                 @endisset
                 <x-ui.button type="submit">Save</x-ui.button>